adding unit test for movie upload and upload validation and refactoring the movie upload

+++ src/controllers/movieUploadController.ts
@@ -1,134 +1,154 @@
 import type { Request, Response } from 'express';
-import { Op } from 'sequelize';
+import { Op, ValidationError, Transaction } from 'sequelize';
 import sequelize from '../sequelize.config';
 import { MovieModel } from '../models';
 
-import { availableGenres, availableCountries, availableMovieLanguages } from '../constants';
-import { UploadUtil, CsvRow, CsvRowWithIndex, UploadError } from '../utils/upload';
+import { DEFAULT_PORT } from '../constants';
+import { UploadUtil, CsvRow, ParsedCsvRowWithIndex, UploadError, parseCsvRow } from '../utils/upload';
 
-async function bulkCreateMovies(rows: CsvRowWithIndex[]) {
-  const transformedRows = rows.map((r) => ({
-    index: r.index,
-    tmdbId: Number(r.tmdb_id),
-    imdbId: r.imdb_id || null,
-    title: r.title,
-    originalTitle: r.original_title,
-    overview: r.overview,
-    runtime: r.runtime,
-    releaseDate: r.release_date,
-    genreIds: Array.from(JSON.parse(rows[0].genres.replaceAll("'", '"'))).map((g) => availableGenres[String(g).toUpperCase()]),
-    originCountryIds: Array.from(JSON.parse(rows[0].countries_of_origin.replaceAll("'", '"'))).map(
-      (c) => availableCountries[String(c).toUpperCase()]
-    ),
-    languageId: availableMovieLanguages[r.language.toUpperCase()],
-    movieStatus: r.movie_status,
-    popularity: r.popularity,
-    budget: r.budget,
-    revenue: r.revenue,
-    ratingAverage: r.rating_average,
-    ratingCount: r.rating_count,
-    posterUrl: r.poster_url,
-    rentalRate: r.rental_rate,
-  }));
+const fieldsForBulkCreate = [
+  'tmdbId',
+  'imdbId',
+  'title',
+  'originalTitle',
+  'overview',
+  'runtime',
+  'releaseDate',
+  'genreIds',
+  'originCountryIds',
+  'languageId',
+  'movieStatus',
+  'popularity',
+  'budget',
+  'revenue',
+  'ratingAverage',
+  'ratingCount',
+  'posterUrl',
+  'rentalRate',
+];
 
-  const newRecords = await sequelize.transaction(async (t) => {
-    const result = await MovieModel.bulkCreate(transformedRows, {
-      fields: [
-        'tmdbId',
-        'imdbId',
-        'title',
-        'originalTitle',
-        'overview',
-        'runtime',
-        'releaseDate',
-        'genreIds',
-        'originCountryIds',
-        'languageId',
-        'movieStatus',
-        'popularity',
-        'budget',
-        'revenue',
-        'ratingAverage',
-        'ratingCount',
-        'posterUrl',
-        'rentalRate',
-      ],
-      ignoreDuplicates: true,
-      logging: false,
-      validate: false,
-      individualHooks: false,
-      returning: ['id', 'tmdbId'],
-      transaction: t,
+async function validateCsvRow(rowNumber: number, row: CsvRow) {
+  const parsedRow = parseCsvRow(row);
+  try {
+    await MovieModel.build(parsedRow, { isNewRecord: true }).validate();
+    // Check for duplicate tmdb id
+    const existingTmdbIdCount = await MovieModel.count({
+      where: {
+        tmdbId: row.tmdb_id,
+      },
     });
-    return result.map((res) => ({
-      id: Number(res.getDataValue('id')),
-      tmdbId: Number(res.getDataValue('tmdbId')),
-      isDuplicate: !res.isNewRecord,
-    }));
-  });
-
-  const newMovieIds = newRecords.map((rec) => {
-    const csvRow = transformedRows.find((r) => r.tmdbId === rec.tmdbId);
-
-    if (!csvRow) {
-      throw new Error(`Error getting the row from csv for tmdb id ${rec.tmdbId}`);
+    if (existingTmdbIdCount > 0) {
+      throw new UploadError('DuplicateTmdbId', rowNumber, `The tmdb_id ${row.tmdb_id} already exist`);
     }
-    if (rec.isDuplicate) {
-      return { rowIndex: csvRow.index, movieId: rec.id === 0 ? null : rec.id, isDuplicate: true };
+    // Check for duplicate imdb id
+    const existingImdbIdCount = await MovieModel.count({
+      where: {
+        imdbId: row.imdb_id,
+      },
+    });
+    if (existingImdbIdCount > 0) {
+      throw new UploadError('DuplicateImdbId', rowNumber, `The imdb_id ${row.imdb_id} already exist`);
     }
-    return { rowIndex: csvRow.index, movieId: rec.id === 0 ? null : rec.id };
-  });
+    return { isValid: true, reasons: [] };
+  } catch (err) {
+    if (err instanceof ValidationError) {
+      const messages = err.errors.map((error) => {
+        if (error.path === 'genreIds') {
+          return `(genres: ${row.genres}) ${error.message}`;
+        } else if (error.path === 'originCountryIds') {
+          return `(countries_of_origin: ${row.countries_of_origin}) ${error.message}`;
+        } else {
+          return `(${error.path}: ${error.value}) ${error.message}`;
+        }
+      });
+      return { isValid: false, reasons: messages };
+    }
 
-  return newMovieIds;
+    if (err instanceof UploadError) {
+      const messages = [err.message];
+      return { isValid: false, reasons: messages };
+    }
+
+    return { isValid: false, reasons: [(err as Error).message] };
+  }
 }
 
 export const trackUpload = async (req: Request, res: Response) => {
+  // res.setHeader('Content-Type', 'text/event-stream');
+  // res.setHeader('Cache-Control', 'no-cache');
+  // res.setHeader('Connection', 'keep-alive');
+  // res.flushHeaders();
+  // try {
+  //   const failedRows: { status: string; data: { rowIndex: number; errors: string[] } }[] = [];
+  //   const { totalRows, processedRows, invalidRows } = await UploadUtil.process(
+  //     3,
+  //     validateCsvRow,
+  //     async (rows: ParsedCsvRowWithIndex[], totalRows: number) => {
+  //       const newMoviesId = await bulkCreateMovies(rows);
+  //       res.write(`data: ${JSON.stringify({ status: 'processing', data: { totalRows, data: newMoviesId } })}\n\n`);
+  //       // await new Promise((resolve) => setTimeout(resolve, 1000));
+  //     },
+  //     (rowIndex: number, errors: string[]) => {
+  //       failedRows.push({ status: 'error', data: { rowIndex, errors } });
+  //       res.write(`data: ${JSON.stringify({ status: 'error', data: { rowIndex, errors } })}\n\n`);
+  //     }
+  //   );
+  //   res.write(
+  //     `data: ${JSON.stringify({
+  //       status: 'done',
+  //       data: {
+  //         totalRows,
+  //         processedRows,
+  //         invalidRows,
+  //         failed: failedRows.length,
+  //         success: processedRows.length,
+  //         failedRows,
+  //       },
+  //     })}\n\n`
+  //   );
+  // } catch (err) {
+  //   const uploadError = err as UploadError;
+  //   res.write(
+  //     `data: ${JSON.stringify({ status: 'api-error', data: { rowIndex: uploadError.rowIndex, message: uploadError.message } })}\n\n`
+  //   );
+  // } finally {
+  //   UploadUtil.deleteFile();
+  //   UploadUtil.cleanUp();
+  //   res.end();
+  // }
+  // req.on('close', () => {
+  //   res.end();
+  // });
+};
+
+export const trackUploadValidation = async (req: Request, res: Response) => {
+  let totalRows = 0;
+  let validRowsCount = 0;
+  let invalidRowsCount = 0;
+
   res.setHeader('Content-Type', 'text/event-stream');
   res.setHeader('Cache-Control', 'no-cache');
   res.setHeader('Connection', 'keep-alive');
   res.flushHeaders();
 
-  try {
-    const failedRows: { status: string; data: { rowIndex: number; errors: string[] } }[] = [];
-    const { totalRows, processedRows, invalidRows } = await UploadUtil.process(
-      3,
-      async (rows: CsvRowWithIndex[], totalRows: number) => {
-        const newMovieIds = await bulkCreateMovies(rows);
-        res.write(`data: ${JSON.stringify({ status: 'processing', data: { totalRows, data: newMovieIds } })}\n\n`);
-        // await new Promise((resolve) => setTimeout(resolve, 1000));
-      },
-      (rowIndex: number, errors: string[]) => {
-        failedRows.push({ status: 'error', data: { rowIndex, errors } });
-        res.write(`data: ${JSON.stringify({ status: 'error', data: { rowIndex, errors } })}\n\n`);
-      }
-    );
+  await UploadUtil.validate(validateCsvRow, (rowNumber: number, isValid: boolean, reasons: string[]) => {
+    if (isValid) {
+      validRowsCount++;
+    } else {
+      invalidRowsCount++;
+    }
+    totalRows++;
     res.write(
-      `data: ${JSON.stringify({
-        status: 'done',
-        data: {
-          totalRows,
-          processedRows,
-          invalidRows,
-          failed: failedRows.length,
-          success: processedRows.length,
-          failedRows,
-        },
-      })}\n\n`
+      `data: ${JSON.stringify({ status: 'processing', index: rowNumber, isValid, ...(reasons.length > 0 && { reasons }) })}\n\n`
     );
-  } catch (err) {
-    const uploadError = err as UploadError;
-    res.write(
-      `data: ${JSON.stringify({ status: 'api-error', data: { rowIndex: uploadError.rowIndex, message: uploadError.message } })}\n\n`
-    );
-  } finally {
-    UploadUtil.deleteFile();
-    UploadUtil.cleanUp();
-    res.end();
-  }
-
-  req.on('close', () => {
-    res.end();
+    res.flush();
   });
+  res.write(`data: ${JSON.stringify({ status: 'done', totalRows, validRowsCount, invalidRowsCount })}\n\n`);
+  res.flush();
+
+  UploadUtil.deleteFile();
+  UploadUtil.cleanUp();
+  res.end();
 };
 
 export const uploadMovies = async (req: Request, res: Response) => {
@@ -141,32 +161,48 @@ export const uploadMovies = async (req: Request, res: Response) => {
   const enable_tracking = 'enable_tracking' in req.query ? req.query.enable_tracking : undefined;
   const enableTracking =
     enable_tracking !== undefined ? ['', 'yes', 'true'].includes(enable_tracking.toString().toLowerCase().trim()) : false;
+  const stop_on_error = 'stop_on_error' in req.query ? req.query.stop_on_error : undefined;
+  const stopOnError =
+    stop_on_error !== undefined ? ['', 'yes', 'true'].includes(stop_on_error.toString().toLowerCase().trim()) : false;
 
   UploadUtil.initiate(filePath);
 
   if (!enableTracking) {
-    // Proceed with processing of csv file and send resposne on completion
+    let t: Transaction | undefined = undefined;
+    if (stopOnError) {
+      t = await sequelize.transaction();
+    }
     try {
-      const { totalRows, processedRows, invalidRows } = await UploadUtil.process(
-        5,
-        async (rows: CsvRowWithIndex[]) => {
-          await new Promise((resolve) => setTimeout(resolve, 1000));
+      const { totalRows, successRows, failedRows } = await UploadUtil.process(
+        async (row: ParsedCsvRowWithIndex): Promise<{ id: number; tmdbId: number }> => {
+          const movieInstance = await MovieModel.create(row, {
+            fields: fieldsForBulkCreate,
+            ignoreDuplicates: false,
+            validate: true,
+            ...(t && { transaction: t }),
+          });
+          return { id: Number(movieInstance.getDataValue('id')), tmdbId: movieInstance.getDataValue('tmdbId') };
         },
-        (rowIndex: number, errors: string[]) => {
-          // throw new UploadError('InvalidRow', rowIndex, JSON.stringify(errors));
-        }
+        undefined,
+        stopOnError
+          ? (errors: UploadError[]) => {
+              throw errors;
+            }
+          : undefined
       );
+      await t?.commit();
       res.status(201).json({
         totalRows,
-        processedRows,
-        invalidRows,
-        failed: 0,
-        success: 0,
-        failedRows: [{ rowIndex: 0, errorMessage: '' }],
+        successRows,
+        failedRows,
       });
     } catch (err: unknown) {
-      const uploadError = err as UploadError;
-      res.status(500).json({ rowIndex: uploadError.rowIndex, message: uploadError.message });
+      await t?.rollback();
+      if (Array.isArray(err) && err.every((e) => e instanceof UploadError)) {
+        res.status(400).json(err);
+      } else {
+        res.status(500).json(err);
+      }
     }
   } else {
     // Do not process the csv and instead wait for the tracking url to be triggered by client
@@ -182,63 +218,35 @@ export const validateUpload = async (req: Request, res: Response) => {
     return;
   }
 
-  res.setHeader('Content-Type', 'text/event-stream');
-  res.setHeader('Cache-Control', 'no-cache');
-  res.setHeader('Connection', 'keep-alive');
-  res.flushHeaders();
+  const enable_tracking = 'enable_tracking' in req.query ? req.query.enable_tracking : undefined;
+  const enableTracking =
+    enable_tracking !== undefined ? ['', 'yes', 'true'].includes(enable_tracking.toString().toLowerCase().trim()) : false;
 
   UploadUtil.initiate(filePath);
 
-  const tmdbIds = new Set<number>();
-  const imdbIds = new Set<string>();
   let totalRows = 0;
-  let validRowsCount = 0;
   let invalidRowsCount = 0;
+  const invalidRows: { rowNumber: number; reasons: string[] }[] = [];
 
-  await UploadUtil.validate((rowNumber: number, rowData: CsvRow, isValid: boolean, reason?: string) => {
-    if (!isNaN(Number(rowData.tmdb_id)) && Number(rowData.tmdb_id) !== 0) {
-      tmdbIds.add(Number(rowData.tmdb_id));
+  if (!enableTracking) {
+    try {
+      await UploadUtil.validate(validateCsvRow, (rowNumber: number, isValid: boolean, reasons: string[]) => {
+        if (!isValid) {
+          invalidRowsCount++;
+          invalidRows.push({ rowNumber, reasons });
+        }
+        totalRows++;
+      });
+    } finally {
+      UploadUtil.deleteFile();
+      UploadUtil.cleanUp();
     }
-    if (rowData.imdb_id.trim() !== '') {
-      imdbIds.add(rowData.imdb_id);
-    }
-    if (isValid) {
-      validRowsCount++;
-    } else {
-      invalidRowsCount++;
-    }
-    totalRows++;
-    res.write(`data: ${JSON.stringify({ status: 'processing', index: rowNumber, isValid, reason })}\n\n`);
-  });
 
-  const moviesWithDuplicateTmdbIds = await MovieModel.findAll({
-    attributes: ['tmdbId'],
-    where: {
-      tmdbId: {
-        [Op.in]: Array.from(tmdbIds),
-      },
-    },
-  });
-
-  const moviesWithDuplicateImdbIds = await MovieModel.findAll({
-    attributes: ['imdbId'],
-    where: {
-      imdbId: {
-        [Op.in]: Array.from(imdbIds),
-      },
-    },
-  });
-
-  const duplicateTmdbIds = moviesWithDuplicateTmdbIds.map((m) => m.getDataValue('tmdbId'));
-  const duplicateImdbIds = moviesWithDuplicateImdbIds.map((m) => m.getDataValue('imdbId'));
-
-  res.write(
-    `data: ${JSON.stringify({ status: 'done', totalRows, validRowsCount, invalidRowsCount, existingTmdbIds: duplicateTmdbIds, existingImdbIds: duplicateImdbIds })}\n\n`
-  );
-
-  UploadUtil.deleteFile();
-  UploadUtil.cleanUp();
-  res.end();
+    res.status(200).json({ totalRows, invalidRows });
+    return;
+  } else {
+    res.status(202).json({ trackingUrl: `${req.protocol}://localhost:${DEFAULT_PORT}/api/movies/upload/track_validation` });
+  }
 };
 
 export const uploadMoviesUsingSocket = async (req: Request, res: Response) => {
